refactor(connector): the file types are published, and the core aliases them
A file handed over by an extension unit and one handed over by a core
connector must be the same type, or the bounded/renamed/content-typed
guarantee is re-decided per producer. ports/jurisdiction already made this
move for the pack contract; this is the same move for attachments.
